Printing of usage does not belong to library
How the program is used is to be defined by the program, not the
library.  Printing of usage has been removed from utils.h and utils.c
under libgzmulti/.

Regardless, if one is to compare output with ls(1), man(1), parted(8),
and git(1), it is noted that the text:

- should not have started with a newline
- should have had a colon after the word "Usage"

Also, program name probably should not be hard-coded.