fcopy - buffer mistyped
size_t is intended to represent how big something is, not the actual
data.  Therefore, declaring the read/write buffer in the fcopy utility
Afunction as size_t is inappropriate.

Also, renamed 'size' to 'count' to be consistent with semantics of
fread(3) and read(2) and their write counterparts, and swapped
parameters 2 and 3 in the calls to fread and fwrite.

Also, rewrote the comment block with the description of what the fcopy
function does, as the original text described the function as copying
compressed data, while it is irrelevant to the function whether or not
the data is compressed.