fsize loses initial file position
According to rewind(3), the function is equivalent to a seek to position
0.  This is apparently not the intention.  The function has been
modified to save the initial file position first and finally restore it
before returning.

P.S. The fsize function is not invoked anywhere in the code so far.  It
is perhaps not the best practice to rely on file size to make decisions,
as the file may be growing in certain situations as it is being
processed.  It may be reasonable to drop this function.