Printing of errors not responsibility of library
In the gzunpack function, if opening the file fails, Z_ERRNO should be
returned, and the caller should be the one responsible for communicating
the situation to the user.

Also, please note that the text of the error message in case the open
failed inaccurately reported not being able to "read" rather than "open"
the file.